+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## X.Y.Z 2022-XX-XX
 ### PaymentSheet
 * [Fixed] PaymentSheet now uses configuration.apiClient for Apple Pay instead of always using STPAPIClient.shared
-
+* [Fixed] Fixed Mac Catalyst support in Xcode 14. [#2001](https://github.com/stripe/stripe-ios/issues/2001)
 ## 22.7.0 2022-08-15
 ### PaymentSheet
 * [Fixed] Fixed a layout issue on iPad.